Rename `doc` module to `piece`
